# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "JACS Studio API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/telemetry/logs": {
      "post": {
        "tags": [
          "telemetry"
        ],
        "summary": "Ingest",
        "operationId": "ingest_api_v1_telemetry_logs_post",
        "parameters": [
          {
            "name": "X-Telemetry-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Telemetry-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelemetryEvent"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "telemetry"
        ],
        "summary": "List Logs",
        "operationId": "list_logs_api_v1_telemetry_logs_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Severity"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 200,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "telemetry"
        ],
        "summary": "Clear All Logs",
        "operationId": "clear_all_logs_api_v1_telemetry_logs_delete",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/v1/telemetry/logs/{log_id}": {
      "delete": {
        "tags": [
          "telemetry"
        ],
        "summary": "Delete Log",
        "operationId": "delete_log_api_v1_telemetry_logs__log_id__delete",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "log_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Log Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/telemetry/logs/manual": {
      "post": {
        "tags": [
          "telemetry"
        ],
        "summary": "Create Manual Log",
        "operationId": "create_manual_log_api_v1_telemetry_logs_manual_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelemetryEvent"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/telemetry/audit": {
      "get": {
        "tags": [
          "telemetry"
        ],
        "summary": "List Audit Logs",
        "operationId": "list_audit_logs_api_v1_telemetry_audit_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 200,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/telemetry/api-operations": {
      "get": {
        "tags": [
          "telemetry"
        ],
        "summary": "Get Api Operations Report",
        "description": "Retrieve dynamic 7-day API operations telemetry report with account activity, error logs, request telemetry, and profit finance breakdown based on registered client devices, jobs, gateway logs, and licenses.",
        "operationId": "get_api_operations_report_api_v1_telemetry_api_operations_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/telemetry/global-requests": {
      "get": {
        "tags": [
          "telemetry"
        ],
        "summary": "List Global Requests",
        "description": "Returns 100% REAL live database logs of all AI requests across all providers, models,\nmachines and API Gateway traffic with real latencies, token counts, costs and statuses.",
        "operationId": "list_global_requests_api_v1_telemetry_global_requests_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 500,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Severity": {
        "type": "string",
        "enum": [
          "warning",
          "error",
          "fatal"
        ],
        "title": "Severity"
      },
...
      "TelemetryEvent": {
        "properties": {
          "event_name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Event Name"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "app_version": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "App Version"
          },
          "fingerprint": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Fingerprint"
          },
          "message": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Message"
          },
          "hwid_hash": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Hwid Hash"
          }
        },
        "type": "object",
        "required": [
          "event_name",
          "severity",
          "app_version",
          "fingerprint",
          "message"
        ],
        "title": "TelemetryEvent"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    }
  }
}